refactored procedure calls and added directives
directives will be fully typechecked and executed as soon as they are parsed, similar to a #run in Jai. but here it's a bit more user-defined so you can just add new directives as you wish as long as they conform to the defined interface. (that interface is still being defined atm)