BIG IF TRUE: added facility for node replacement during typechecking
user can now provide a callback for identifier resolution
this callback also provides the option to simply replace the identifier node.
Currently this node replacement facility is only used for this one case, but it affects a decent bit because we now have to pass a **Node to typecheck_node instead of just a *Node. Alternatively, I may change this so that a replacement node is instead returned to the caller explicitly, but atm it is what it is