refactored nodes to remove `next` member
Now all nodes that reference lists of nodes will use an array view.
There's not really any performance benefit to this per se, but it makes some code simpler to reason about and will make it easier to replace nodes in the future.
The current pattern we use while parsing is to build the list of nodes in the temporary allocator and then copy the completed list into the script's pool. This should prevent some wasted space in the pool and fragmentation of nodes.