part-way through removing dependency on dyncall
still need to fix directives up, but then we should be at least minimally operational without the need for dyncall. Non-dyncall mode will still necessarily be a bit more restricted, since we are forced to generate any wrapper procs at compile-time. This means we can't call procedures returned by arbitrary expressions, only procedures which were exposed directly to the script with register_procedure. This is quite a big drawback, since it means scripts can now never make use of function pointers. But oh well, user beware I guess. For the record, dyncall will still be supported, it's just that it's going to be opt-in rather than a requirement.